LayoutTests/fast/text/international/system-language/navigator-language/navigator-language-pt-BR-expected.txt

PASS navigator.language is "pt"
PASS successfullyParsed is true

TEST COMPLETE
